search-tutorial/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
blinker==1.7.0
boto3==1.28.73
botocore==1.31.73
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
dataclasses-json==0.6.1
elastic-transport==8.10.0
elasticsearch==8.11.1
exceptiongroup==1.1.1
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
glasto==0.0.4
h11==0.14.0
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
langchain==0.0.325
langsmith==0.0.53
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
outcome==1.2.0
packaging==23.1
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.1
pyarrow==14.0.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.17.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.28.2
rich==13.7.0
rpds-py==0.13.1
s3transfer==0.7.0
safetensors==0.4.0
selenium==4.9.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.22
streamlit==1.28.2
tenacity==8.2.3
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
transformers==4.35.2
trio==0.22.0
trio-websocket==0.10.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.15
validators==0.22.0
webdriver-manager==3.8.6
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
